+++ apps/nextjs-website/src/lib/fetch/__tests__/navigation.test.ts
@@ -96,7 +96,6 @@ describe('getNavigation', () => {
         path: '/',
         title: 'Homepage',
         related: {
-          slug: 'homepage',
           sections: [
             {
               __component: 'sections.hero',
