+++ apps/nextjs-website/src/lib/fetch/navigation.ts
@@ -1,134 +1,9 @@
 import * as t from 'io-ts';
 import { extractFromResponse } from './extractFromResponse';
+import { PageSectionCodec } from './types/page';
 import { AppEnv } from '@/AppEnv';
 
 // Codec
-const HeroSectionCodec = t.strict({
-  __component: t.literal('sections.hero'),
-  title: t.string,
-  subtitle: t.union([t.string, t.null]),
-  theme: t.union([t.literal('light'), t.literal('dark')]),
-  inverse: t.boolean,
-  size: t.keyof({
-    small: null,
-    big: null,
-  }),
-  useHoverlay: t.boolean,
-  sectionID: t.union([t.string, t.null]),
-  image: t.union([
-    t.strict({
-      alternativeText: t.union([t.string, t.null]),
-      url: t.string,
-    }),
-    t.null,
-  ]),
-  background: t.union([
-    t.strict({
-      alternativeText: t.union([t.string, t.null]),
-      url: t.string,
-    }),
-    t.null,
-  ]),
-  ctaButtons: t.array(
-    // TODO: Replace with CTAButtonSchema when merged
-    t.intersection([
-      t.type({
-        text: t.string,
-        href: t.string,
-        variant: t.keyof({
-          text: null,
-          outlined: null,
-          contained: null,
-        }),
-        color: t.keyof({
-          inherit: null,
-          primary: null,
-          secondary: null,
-          success: null,
-          error: null,
-          info: null,
-          warning: null,
-        }),
-      }),
-      t.partial({
-        icon: t.union([t.string, t.null]),
-        size: t.keyof({
-          small: null,
-          medium: null,
-          large: null,
-        }),
-      }),
-    ])
-  ),
-});
-
-const EditorialSectionCodec = t.strict({
-  __component: t.literal('sections.editorial'),
-  title: t.string,
-  eyelet: t.union([t.string, t.null]),
-  body: t.string,
-  theme: t.union([t.literal('light'), t.literal('dark')]),
-  pattern: t.keyof({
-    // Move to separate Codec?
-    none: null,
-    dots: null,
-    solid: null,
-  }),
-  width: t.keyof({
-    // Move to separate Codec?
-    standard: null,
-    wide: null,
-    center: null,
-  }),
-  reversed: t.boolean,
-  sectionID: t.union([t.string, t.null]),
-  image: t.union([
-    t.strict({
-      alternativeText: t.union([t.string, t.null]),
-      url: t.string,
-    }),
-    t.null,
-  ]),
-  ctaButtons: t.array(
-    // TODO: Replace with CTAButtonSchema when merged
-    t.intersection([
-      t.type({
-        text: t.string,
-        href: t.string,
-        variant: t.keyof({
-          text: null,
-          outlined: null,
-          contained: null,
-        }),
-        color: t.keyof({
-          inherit: null,
-          primary: null,
-          secondary: null,
-          success: null,
-          error: null,
-          info: null,
-          warning: null,
-        }),
-      }),
-      t.partial({
-        icon: t.union([t.string, t.null]),
-        size: t.keyof({
-          small: null,
-          medium: null,
-          large: null,
-        }),
-      }),
-    ])
-  ),
-});
-
-const PageSectionCodec = t.union([HeroSectionCodec, EditorialSectionCodec]);
-
-const PageDataCodec = t.strict({
-  slug: t.string,
-  sections: t.array(PageSectionCodec),
-});
-
 const ParentCodec = t.strict({
   order: t.number,
   id: t.number,
@@ -141,14 +16,15 @@ const NavItemCodec = t.intersection([
   ParentCodec,
   t.strict({
     parent: t.union([ParentCodec, t.null]),
-    related: PageDataCodec,
+    related: t.strict({
+      sections: t.array(PageSectionCodec),
+    }),
   }),
 ]);
 const NavigationCodec = t.readonlyArray(NavItemCodec);
 
 // Types
 export type Navigation = t.TypeOf<typeof NavigationCodec>;
-export type PageSection = t.TypeOf<typeof PageSectionCodec>;
 
 export const getNavigation = (
   menuName: string,
